Move exception_list to executors module
