+++ libs/algorithms/include/hpx/parallel/algorithms/detail/dispatch.hpp
@@ -15,8 +15,8 @@
 #include <hpx/type_support/decay.hpp>
 
 #include <hpx/datastructures/tuple.hpp>
-#include <hpx/execution/exception_list.hpp>
 #include <hpx/execution/executors/execution.hpp>
+#include <hpx/executors/exception_list.hpp>
 #include <hpx/executors/execution_policy.hpp>
 #include <hpx/parallel/util/detail/algorithm_result.hpp>
 #include <hpx/parallel/util/detail/scoped_executor_parameters.hpp>
