+++ libs/algorithms/include/hpx/parallel/task_block.hpp
@@ -20,8 +20,8 @@
 #include <hpx/traits/is_future.hpp>
 #include <hpx/type_support/decay.hpp>
 
-#include <hpx/execution/exception_list.hpp>
 #include <hpx/execution/executors/execution.hpp>
+#include <hpx/executors/exception_list.hpp>
 #include <hpx/executors/execution_policy.hpp>
 #include <hpx/parallel/util/detail/algorithm_result.hpp>
