+++ libs/executors/include_compatibility/hpx/parallel/exception_list.hpp
@@ -5,15 +5,15 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/config.hpp>
-#include <hpx/execution/config/defines.hpp>
-#include <hpx/execution/exception_list.hpp>
+#include <hpx/executors/config/defines.hpp>
+#include <hpx/executors/exception_list.hpp>
 
-#if defined(HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_EXECUTORS_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/parallel/exception_list.hpp is deprecated, \
-    please include hpx/execution/exception_list.hpp instead")
+    please include hpx/executors/exception_list.hpp instead")
 #else
 #warning "The header hpx/parallel/exception_list.hpp is deprecated, \
-    please include hpx/execution/exception_list.hpp instead"
+    please include hpx/executors/exception_list.hpp instead"
 #endif
 #endif
